+++ Sources/UInt8Util/UInt8+String.swift
@@ -8,12 +8,6 @@ extension String {
     self.init(bytes: asciiValues, encoding: .ascii)!
   }
 
-  @available(*, deprecated, renamed: "init(asciiValues:)")
-  @inlinable
-  @inline(__always)
-  public init<S>(ascii: S) where S: Sequence, S.Element == UInt8 {
-    self.init(bytes: ascii, encoding: .ascii)!
-  }
   
   @inlinable
   @inline(__always)